Fix unique_ptr swap implementation
The unique_ptr internal swap function tries to swap resources with
another unique_ptr but the compiler resolved the scope of the swap call
to be within the class definition for which there isn't a function that
can handle the swap.

This commit adds a using std::swap statement to the unique_ptr swap
function which resolves the issue by pulling std::swap into the scope of
that function so the compiler can select the right function to use.

Change-Id: Ia27adc141c1941d1b6b02705172ccf35156d7b56
RTC:300781
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131427
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>