IO SCM chip deconfiguring post merge cleanup
Simplified the getAssociatedIoScmChip method to not build up
a mapping but inspect each PROC looking for pairs based
on the given target location code.

Also redefined what it means to have a DCM.  Now if any
pair is found, then the system is considered a DCM as opposed
to that all PROCs must be paired.

Change-Id: I879ca5db88de525e9adb46eed3968e22a5a1594f
RTC: 301401
CQ: SW546709
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131450
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>