Define the SYS_CLOCK_INTEGRATED_SPARES attribute
This commit defines the SYS_CLOCK_INTEGRATED_SPARES attribute. This
attribute will be set to 1 for systems that have clocks that can be
configured as integrated spares (BMC systems with multiple clocks),
and 0 otherwise. This attribute will be set appropriately in the MRW
for each system, and will be used by PRD to determine the service
strategy for clock error logs.

Change-Id: Ibb5c01cde4d0aa64a413fb1939cbe143326b1613
RTC: 295271
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131625
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>