Add hash of image to image ID string
The PNOR code image that hostboot generates (a combination of hbicore.bin
and hbicore_extended.bin) contains an image ID that indicates what git
commit the components were built with. If the git working tree is
dirty, however, the image ID only includes the word "dirty," but
doesn't provide a way to identify what changes were made to the
working tree.

This commit introduces a change to add a partial SHA1 checksum to the
long image ID so that the exact binary that is running on the system
can be identified. This is useful for things like identifying patches on
a system, and for tools that need to verify binary compatibility with
the installed Hostboot images (such as DCE).

The SHA1 is a hash of hbicore.bin followed by
hbicore_extended.bin (test versions use the appropriate binaries).

For example, in a patched system, the previous message showed:

  Welcome to Hostboot branched-fw1020-24-g829e8ca-dirty/hbicore.bin/zachc...

The new banner shows (backslash is line continuation for commit message
line width restrictions):

  Welcome to Hostboot
   \ branched-fw1020-24-g829e8ca-dirty-sha1:471cad31/hbicore.bin/zachc...

Change-Id: I3e5bda5bff384afa97ce2d1ac10ba737018841e4
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131706
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>