Fix FCO deconfig actions to deconfigure FCs with no COREs
The new FCO algorithm assumed that child roll up would handle deconfig
of the FCs when all of its children were deconfigured. This was not the
case leading to watchdog timeouts whenever FCO was applied to a system.
These timeouts occurred when jumping to PHYP where it would wait forever
to be told about the FCs that had be deconfigured by FCO incompletely.

Change-Id: Id50f3ec7430aa6878290e59f639410f76a1df255
CQ:SW550667
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131700
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>