Handle deconfigured boot core
There is a scenario where Hostboot and BMC can disagree on which
cores are functional.  In this case the BMC could direct the SBE
to boot Hostboot onto a core that Hostboot believes is not
functional.  This will cause an IPL failure that leads to an
infinite loop because the two sides will never quite agree.

If there truly aren't any viable cores to use we will fail but in
most cases there are plenty of cores we could use instead.  The
solution is for Hostboot to explicitly for a reconfig loop if
we find any viable cores to use instead.  On the reboot, BMC will
consume the HWAS_STATE that we sync down and therefore choose a
core that Hostboot believes is valid.  The list of viable cores
would be any non-ECO core that is either functional or deconfigured
due to Field Core Override.

Change-Id: I2ccd9e66d95e2c4cc99d55f5c7f722f25e663e68
CQ: SW550607
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131777
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>