PRD: Disable predictive clock callouts on Everest
Everest is configured with integrated spare clocks. We don't want to
tell the customer to replace the parts until both clocks are bad. So PRD
will explicitly prevent a predictive callout for RCS and PLL unlock
attentions so that no service is required until all clocks have failed.
In which case, the checkstop analysis code will make the expected
callouts.

Change-Id: Ic23afa5eadecfd6f256bd47310a47ab0d7a33ae2
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131761
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Benjamen G Tyner <ben.tyner@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>