Attr Dump: AtrrRp Changes
This commit contains changes to AtrrRp to perform an attribute
dump during HB boot. The attributes are formatted and written
out to a lid file (0x81e00666) via PLDM.

The code in this commit is capable of dumping simple-type,
single-value and array attributes. The values of the attributes
are fetched directly from the memory.

Change-Id: I0c4ebdf75c8dd4e2c6d048d665858df5a9f129c4
RTC: 213109
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131839
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>