Add functions to add arbitrary systems to FCO tests
This commit adds new functions to allow devs to create any system config
given a bit string array. Should a testcase fail as a result of a
chaotically generated system, the output bit string in the traces can be
used as input into the systemConfigurationOptions vector to create a new
test.

This commit also makes various other improvements to the FCO test case.

Change-Id: If0517a635e962acb966c5f761a7c3d8a14951fcf
RTC:300781
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131840
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>