HDAT: Unpopulated DIMM and Processor support
This commit adds all the dimm and processors in to the SLCA table
even though the corresponding slots are empty. This commit fills
the connector PLDM entity info for each of the dimm and processor
slots into the SLCA table. Host can use this for the LED operations
by fetching the PLDM entity information and comparing against the
PDR data.

Change-Id: I041325e9d5004eb3651f2cc1aa1b19d3a34d77fc
CQ: SW551210
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/131934
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>