Pin Idle Task in Page Table
The idle task function is the default task for any idle thread.
However, it is treated like any regular user-space task in terms
of page table management.  This can lead to an extreme slowdown
situation if the page table entry (PTE) for this function happens
to fall out of the page table.  If this happens and we have 100s
of threads active, we can end up in a cycle of PTE misses that
get stuck on a spinlock and then end up retriggering as the PTE
is manipulated.

One possible solution is to ensure that the PTE representing the
idle function is never evicted from the page table.

Change-Id: Ia2aa2b2787b2cbfc197799a037ddd9e3039ceedd
CQ: SW550780
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/132138
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>