PRD: Ignore attentions from EQ_CORE_FIR[60]
The hardware team recently unmasked this bit to enable recovery. This
bit simply indicates there is a core unit checkstop attention somewhere
else in this FIR. This could even be from EQ_CORE_FIR[57] which
indicates a unit checkstop attention came from the other core in the
fused core pair. We will ignore this attention by prioritizing it as the
last possible attention to analyze in the FIR. This should ensure we
see the real unit checkstop attentioni whether it is on this core or on
the other core of the pair.

Change-Id: Icdccd768be8b67dfd46bfcd2b5cf2c849bde8bfe
CQ: SW551584
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/132121
Reviewed-by: Caleb N Palmer <cnpalmer@us.ibm.com>
Reviewed-by: Austin F Cui <austin.cui@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/132172
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>