Remove FCO reference from restrictECunits doxygen
There was a lingering comment that implied restrictECunits handles FCO
which is no longer does. That comment was removed and in the body of the
function a note has been left to tell readers that FCO is no longer a
supported feature of that function.

The FCO support in the function needs to be pulled out still but that
will come later. For now, these changes serve to avoid confusion.

Change-Id: If358f8ce2a3dd36524564a48a8ddb1827d80e93f
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/132440
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Corey V Swenson <cswenson@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>