Reset/clear global variables
When the hostfw_trace.py is called by the peltool script, the global
variables to do not get cleared on subsequent calls to decode a trace
buffer.  This can leave warnings and/or errors behind, if found, for a
buffer which will be propagated to any subsequent buffer decoding.

Resetting the global variables gives a "clean slate" for each buffer decoding
to work with and eliminates polluting the decoded buffer.

This "feature" can be advantageous when reading in the string file into
a hash. Reading the string file into a hash has been determined to take
a considered amount of time relative to the decoding overall, therefore if
we do not have to reload this file for every buffer decoding then
time performance can be improved.  A future commit for this will be submitted.

Change-Id: I5fe0539478d002b8e49045162a3beeb1509e36f0
CQ: SW552597
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/132477
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>
Tested-by: Nick Bofferding <bofferdn@us.ibm.com>