Do not rely on FRU table length in metadata being correct
The PLDM FRU table metadata, which Hostboot fetches from the BMC,
contains the size of the FRU table. Previously, Hostboot was trusting
that this size was accurate and using it to make allocations, but it
is possible for the declared size and the actual size to be
different. This commit changes the decoder to use the actual size of
the PLDM response data in the allocation instead of the size given in
the FRU table metadata, which can't be "wrong."

Change-Id: I118437e572bc98d2c9fdad03f717cc33b566e628
CQ: SW552529
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/132591
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>