Peltool performance enhancements
When the hostfw_trace.py is called by the peltool script, the global
variables to do not get cleared on subsequent calls to decode a trace
buffer.  This "feature" can be advantageous when reading in a string file into
a hash. Reading the string file into a hash has been determined to take
a considered amount of time relative to the decoding overall, therefore if
we do not have to reload this file for every buffer decoding then
time performance can be improved.

This commit adds code to check if the string file has already been encountered
and if so, use the saved data from the file instead of re-processing the file
again.  If the string file has never been encountered before then process the
data within the file and preserve the data for future use.

Did the same for 0x0100.  I create global vars to preserve the symbol
table data once it has been processed.

Change-Id: I22f3a94d4dec6a48cd7b37e59731a5a7c6740dc0
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/132655
Reviewed-by: Deborah L McLemore <debmc@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>