Attr Dump: HBRT
This commit implements changes to be able to use attribute
dump during runtime. Part of these changes is commonization
of attribute dump code between HB and HBRT.

Note that the map that maps attribute hash to its data
needed to be changed to an array because HBRT was too
resource-limited to be able to init a map of this size
at runtime.

Change-Id: I23cbeaec7c93aaae1fa3cdced3bcf10a8f2700a6
RTC: 213109
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/132727
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Mike Baiocchi <baiocchi@ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>