clear/reset global vars in file hostfw_trace.py
When the hostfw_trace.py is called by the peltool script, the global
variables to do not get cleared on subsequent calls to decode a trace
buffer.  This can leave warnings and/or errors behind, if found, for a
buffer which will be propagated to any subsequent buffer decoding.

Resetting the global variables gives a "clean slate" for each buffer decoding
to work with and eliminates polluting the decoded buffer.

Also this commit adds versioning to the peltool script.

Change-Id: I519c29e25f58efa1b3b6e2d9fddb0b71f51fe70c
CQ: SW552597
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/132720
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>