Decrease MCTP packet buffer size to 16KB from 128KB
We have seen cases where the Hostboot memory is too fragmented
to handle mallocs of 20KB at the point where we are reading the
(large) lids from the BMC into memory.  The amount of data that
we allocate for the transfer at any given time is currently set
to ~128KB.  This change drops that to ~16KB so that it fits into
the observed contiguous memory buckets that we seem to have.

Note - There was no observable change in the total IPL time with
this reduction.

Change-Id: I5d20b54963752a548f79e24c8dc049b7979a53fd
CQ: SW550406
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/132725
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>