PRD: update current parser for Line Delete tables
The parser has a few issues. If PPC data is parsed with an x86 errl
tool, the parsed information is garbage. Then, the L3 data is actually
broken because the bit fields for the member and bank are not big
enough. Also, there is no field for the cl half. Did the best we could
with the data we get, but will eventually deprecate the data and move it
to a new user data section. The legacy parser will still remain for
legacy data, if it exists.

Change-Id: I87b9e996797ab6d452799d2d0300cf8a9f4d9baa
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/132832
Reviewed-by: Benjamen G Tyner <ben.tyner@ibm.com>
Reviewed-by: Austin F Cui <austin.cui@ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/132941
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>