Explictly set correct default SCOM_SWITCHES for all procs
At initial boot, we can use XSCOM to scom on the bootproc and
only the fsi2pib engine on the secondary procs.  Prior to this
commit those settings were made by the processMrw script based
on the current boot processor.  This caused a problem in local
simics that uses a hand-coded attribute xml since that logic
wasn't applied.  A better solution is to have the code explicitly
choose the appropriate values based on the live system state.

Change-Id: Ief4ac149de4909d4a8cb35580cd7e6c6638bcd44
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/132818
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Matt Derksen <mderkse1@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>