PRD: Handling for host detected LPC timeout
For reasons not explained yet, hardware will not initiate an LPC timeout
attention via NCU timeout FIR bit as we expected on BMC based machines.
When Hostboot detects an LPC timeout, it will manually set
N1_LOCAL_FIR[61] to force a system checkstop. The service response for
this bit will be to call out the hardware as if there was a hardware
reported LPC timeout. Note that this workaround is only expected on BMC
based machines, but making the same changeto FSP PRD just in case.

Change-Id: Ice656e7b61eb14f4b7d279784570d17132c21af1
CQ: SW552763
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/132880
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Reviewed-by: Benjamen G Tyner <ben.tyner@ibm.com>
Reviewed-by: Austin F Cui <austin.cui@ibm.com>
Reviewed-by: Zane C Shelley <zshelle@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/132942
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>