Improvements to DCE
Updates the DCE compile script and makefile to fix some of the hardcoded
CXXFLAGS. Now DCE will pickup those flags through the usual method that
hostboot uses during compilation. Additionally, this allows users to
specify their own CXXFLAGS if they need to.

Added a dce_rc file that has the env vars DCE uses so that users can
update that file and source it prior to compilation instead of suppling
them on the command line if desired.

Adds some declares for FAPI globals so that users may use FAPI code with
minimal hassle.

Change-Id: Ifb359d500695941a1b6a198dda36904d2d6c9a35
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/132894
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Zachary Clark <zach@ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>