Support BIOS control of max freq setting (for HANA)
The plan for 24core in Everest and Rainier is to have WOF table
with UT at 3.9GHz. For HANA certification, we may need to lower
this frequency to 3.6GHz for noise issue.  Ideally this value
would be flexible so it is implemented as a new PLDM BIOS attr
hb_cap_freq_mhz_request.

Change-Id: Ic7745ea1cc481dc86bd16450ee8c9158b4db3164
RTC:305086
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/132939
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Isaac Salem <isaac.salem@ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Roland Veloz <rveloz@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>