Sort PLDM OCC state sensor/effecter IDs
The BMC does not yet expose the PLDM entity tree to BMC
applications. This means that BMC apps don't have any general-purpose
way to associate PLDM sensor/effecter IDs with targets, because they
can't traverse the entity hierarchy if they need to. This is a problem
for the OCC application, for example, because it must be able to tell
which processor is associated with each OCC state sensor in order to
use them, but it needs access to the entity tree to do this because
the sensor/effecter entity instance IDs are not unique (e.g. four
processors across two DCMs will have instance IDs 0, 1, 0, 1; the
disambiguation must be done by examining the entity container, which
applications can't do without the entity tree).

To fix the problem this for OCC sensors, this commit updates the PDR
generation code to sort the OCCs by the ORDINAL_ID attribute of their
parent processors before creating their sensors and effecters. Since
the values of sensor/effecter IDs increase as they are generated, this
will allow the OCC application to figure out which ID goes with which
processor by sorting the sensors by numeric value, and sorting a list
of processors by the values of their ORDINAL_ID attributes. The nth
element of the sensor/effecter ID list will correspond with the nth
element of the processor list.

CQ: SW553391
Change-Id: I54687265652072696e677320666f722074686520
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/133161
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Chris Cain <cjcain@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>