Update Method To Force a Checkstop on some LPC Errors
This commit updates the method used to force a checkstop on any
LPC Errors that are not masked off such that PRD running on the FSP
or BMC can correctly callout the hardware.

Previously, when LPC errors were seen the EQ_L2_FIR[13] FIR bit
was set across all of the cores in the boot proc.  Now the
'N1_LOCAL_FIR[61] - Host detected LPC timeout' bit will be set to
force a checkstop.

This commit makes this update to both the Hostboot Bootloader (HBBL)
and to the Hostboot Base Image (HBB).

Change-Id: Ifa398e9ead4ef64e6ee5223e648aa398893cb398
CQ:SW552763
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/133164
Reviewed-by: Ilya Smirnov <ismirno@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>