Hostboot getCoreSelect for unicast target
When plat_target_handle_t class is constructed for a unicast target, seed the
coreSelect instance with the proper multicast mask based on the relative
core position in the EQ.

Subsequently when getCoreSelect is run for a target, the proper coreSelect
value will be able to be retrieved.

The root cause of this workaround is due to some HWPs expecting a unicast
target to contain multicast attributes which are not explicitly set for
unicast targets.

Change-Id: I774b5d34096aa13139c8b591dc66fbf0867fe647
CQ:  SW553007
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/133290
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>