Change FCO sorting criteria
Changes the FCO sorting criteria for procs so that if the procs have
equal number of available cores then the proc with the lower HUID will
be placed ahead in the list.

This change is necessary because hostboot std::sort uses an unstable
sorting algorithm and in the case where procs have equal amount of
available cores it makes it difficult to predict which procs will be
assigned cores with low FCO values.

Change-Id: I801e909afa15f55dbbd6ecf9f0a1609e28a073bf
CQ:SW553396
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/133333
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>