Ignore merging preserved attributes on MPIPL in targeting init
There is a sequencing error where during targeting init on an MPIPL,
Hostboot applies preserved attributes from the HBD_RW file residing
on BMC onto the mainstore attribute values carried over from HBRT, which
is wrong.  This happens because the mergeAttributes API reads the
attribute indicating if Hostboot is in an MPIPL before that attribute
has actually been updated.  This commit adds an MPIPL flag to the
mergeAttributes API and modifies targeting init to directly pass that
flag to said API so that it does not need to rely on the stale attribute
and can thus make the right decision.

Change-Id: I7d7788a7771e018c2e0b96e7c25d4e4e35e46b3d
CQ: SW553620
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/133374
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>