Add new LMB size strings to an internal vector of possible values
While new strings were added to supprt the new 1024MB, 2048MB, and
4096 LMB sizes, these strings were not added to the vector
POSSIBLE_HB_MEM_REGION_SIZE_STRINGS that is used to see if there is
a match to the bios attribute sent from the BMC to hostboot.

Fortunately, the strings were still being used correctly to properly
set the ATTR_LMB_SIZE once the match was determined.

Change-Id: I7d3f1aa11477de07ccda677ce23731ff8bc85e0c
RTC:258716
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/134144
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Chen Du <duchen@us.ibm.com>
Reviewed-by: Matthew Raybuck <matthew.raybuck@ibm.com>
Reviewed-by: Nick Bofferding <bofferdn@us.ibm.com>