Updates CCS instructions to dynamically add idles and repeats
Odyssey has 16 bit idle and repeat counters but only 8 bits allocated
for the idle and repeats in the register. As such, there is a padding
field within the register. However, the padding needs to be worked out
based upon the desired idles and repeats.

Explorer has 16 bits allocated for idle and repeat each.

This commit adds an instance variable for idles and repeats in the
CCS instruction. It then adds functions to update the idles and repeats.
It was done in prep for the addition of Odyssey's CCS specialization

ZEN:MST-1676
Change-Id: Ie99cf2d628e309e34839c5a0c46cc2db347cb4a2
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/134242
Tested-by: Jenkins Server <pfd-jenkins+hostboot@us.ibm.com>
Tested-by: FSP CI Jenkins <fsp-CI-jenkins+hostboot@us.ibm.com>
Reviewed-by: Sneha Kadam <sneha.kadam1@ibm.com>
Dev-Ready: STEPHEN GLANCY <sglancy@us.ibm.com>
Reviewed-by: Louis Stermole <stermole@us.ibm.com>
Tested-by: Hostboot CI <hostboot-ci+hostboot@us.ibm.com>
Reviewed-by: Edgar Cordero <ecordero@us.ibm.com>
Reviewed-on: http://rchgit01.rchland.ibm.com/gerrit1/134350
Tested-by: Jenkins OP Build CI <op-jenkins+hostboot@us.ibm.com>
Tested-by: Jenkins Combined Simics CI <combined-simics-ci+hostboot@us.ibm.com>
Tested-by: Jenkins OP HW <op-hw-jenkins+hostboot@us.ibm.com>
Reviewed-by: Daniel M Crowell <dcrowell@us.ibm.com>